refactor: 페이지 라우팅 로딩 hook 구현, Spinner 스타일 조정
